React Udemy Project 2 - Bug Fix
Button disabling functionality starts to work only when the user has some interaction with the input field.
Initially when the app was rendered, even though the input field is empty, the button was clickable and this allows the form to submit an empty string.

Fix - Modified the logic for disabling the button, when the input field is empty, the button has to be disabled.